Move around some storage classes to make the proper metrics methods available to workers.
